Lazily initialize settings on the client
This ensures that initial state and settings are parsed by the client before the
react app is initialized. Previously under certain timing conditions (for example
when there was lots of data in the initial data script tag) the react app got
initialized with default settings not the correct values from SSR. This can cause
weird issues like the frontend attempting to do api calls to http://django.

This is resolved by delaying initializing react app until the initial data
script tag is parsed. In case client.js is parsed after the initial data node
then it just renders the react app. But if it is parsed before it just exposes
global maybeInitReactApp method. This global is called at the end of initial data
script node. Double initialization is prevented by a global `__app_initialized__`.

The settings themselves have been turned into a module variable which gets updated by
a call to loadSettings. The SETTINGS object is exported trough an ES6 Proxy which allows
us to update the contents of the exported value during runtime.