Fixed AWS S3 access denied error on new projects.
Introduced a new boolean configuration variable `s3_media_bucket_is_public` in the top level variables.tf file.
This var is passed down to the `s3_media` module.

This variable controls some of the `aws_s3_bucket_public_access_block` settings.

Set it to false to be consistent with the Django settings and the usage pattern.
Otherwise, AWS blocks all public access to the bucket.

Added a comment to PrivateMediaStorage.