Unify AWS S3 region throughout the project
Previously the Terraform scripts, which are creating the buckets, we're creating S3 buckets in "eu-north-1" region, but Django settings we're trying to access the buckets in "eu-central-1" region.

As a result, when trying to access the uploaded files through Django you would receive a PermanentRedirectError because the regions we're not matching.