Update ansible dependencies
Fix issue with POSTGRES_USER/POSTGRES_PASSWORD not being set to correct values locally (for example during mirroring
or when first creating a project). The issue was due to the syntax of defaults causing the resulting value to be
`'repo_name'` (with the quotes included). Fixed the syntax and also ensured that the .env file is loaded for all
service containers too (to allow overwriting the env vars).