Update DJANGO_IMAGE_NAME to use dash instead of underscore
Docker changed the behaviour of the container naming to use `-`. Originally we
believed that this change was reverted in 2.9.0 as it caused a breaking change.
But it turns out instead they reverted it only for related resources such as
networks, volumes, secrets and config.

This means that going forward the container names use dashes (`-`).

Note: Also updated the poetry wrapper image name to match this new convention.

Refs:

- https://docs.docker.com/compose/release-notes/#290
- https://github.com/docker/compose/pull/9700
- https://github.com/docker/compose/pull/9700#issuecomment-1206382112
- https://github.com/compose-spec/compose-go/pull/297
- Internal ref: https://thorgate.slack.com/archives/C02AMG235/p1659433239011469

Additional changes in the PR:

- Update poetry installation command
  - ref: https://python-poetry.org/docs/master/#installing-with-the-official-installer
- Remove pypi docker compose in ci image (docker-compose now comes from inside docker image anyway)