Make it possible to use all types of urls commonly set as gitlab_repo_url
`gitlab_repo_url` expected the website url of the repository and used python
to convert this to a git url or docker registry urls. This works nicely when
people use the "correct" url but its confusing which one is correct.

This PR changes the filters to also work for other commonly used url formats in
that variable:

- git ssh url
- git http/https url
- various urls to git repo without protocol

I also added some basic tests to the filters to ensure the expected functionality
does not break accidentally in the future. Additionally the variable documentation
was updated to reflect its usage at this point.